+++ arch/msp430/msp430_alu.h
@@ -16,7 +16,7 @@ typedef int16_t mcu_register_t;
 struct msp430_alu_t {
   mcu_register_t   regs[MCU_REGISTERS];
   int              interrupts;
-  int              interrupt_vector;
+  uint16_t         interrupt_vector;
   uint64_t         insn_counter;
   uint64_t         cycle_counter;
   uint32_t         signal;
